requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Graph Neural Networks
torch-geometric>=2.3.0
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv

# Computer Vision
opencv-python>=4.8.0
imageio>=2.31.0
Pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
networkx>=3.1

# Progress bars and utilities
tqdm>=4.65.0

# TensorBoard
tensorboard>=2.13.0

# Object Detection (GroundingDINO)
groundingdino-py>=0.4.0

# Segmentation (SAM 2)
# Install via: pip install 'git+https://github.com/facebookresearch/segment-anything-2.git'

# Ego4D (for data download)
ego4d>=1.6.0

# AWS CLI (for Ego4D data)
awscli>=1.29.0

# Hugging Face (for model downloads)
huggingface-hub>=0.16.0
transformers>=4.30.0

